Add adapter-base modules
Added the adapter-base, adapter-base-spring and adapter-base-quarkus
modules which are supposed to contain classes used for building protocol
adapters based on Spring and/or Quarkus.

Initially, the base classes for implementing adapters and the
Spring/Quarkus specific configuration classes have been moved to the new
modules.
