+++ adapters/base-quarkus/pom.xml
@@ -50,6 +50,10 @@
   <dependencies>
     <dependency>
       <groupId>org.eclipse.hono</groupId>
+      <artifactId>hono-adapter-base-quarkus</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.hono</groupId>
       <artifactId>hono-service-base-quarkus</artifactId>
     </dependency>
     <dependency>
