+++ adapters/base-spring/pom.xml
@@ -28,6 +28,10 @@
   <dependencies>
     <dependency>
       <groupId>org.eclipse.hono</groupId>
+      <artifactId>hono-adapter-base-spring</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.hono</groupId>
       <artifactId>hono-service-base-spring</artifactId>
     </dependency>
     <dependency>
