+++ bom/pom.xml
@@ -149,6 +149,21 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.hono</groupId>
+        <artifactId>hono-adapter-base</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.hono</groupId>
+        <artifactId>hono-adapter-base-quarkus</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.hono</groupId>
+        <artifactId>hono-adapter-base-spring</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.hono</groupId>
         <artifactId>hono-service-base</artifactId>
         <version>${project.version}</version>
       </dependency>
@@ -848,7 +863,7 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.hono</groupId>
-        <artifactId>service-base-test-utils</artifactId>
+        <artifactId>adapter-base-test-utils</artifactId>
         <version>${project.version}</version>
         <scope>test</scope>
       </dependency>
