+++ pom.xml
@@ -217,12 +217,14 @@
   </properties>
 
   <modules>
+    <module>adapter-base</module>
+    <module>adapter-base-quarkus</module>
+    <module>adapter-base-spring</module>
     <module>adapters</module>
     <module>bom</module>
     <module>core</module>
     <module>cli</module>
     <module>client</module>
-    
     <module>client-device-connection-infinispan</module>
     <module>clients</module>
     <module>demo-certs</module>
