+++ services/pom.xml
@@ -95,10 +95,6 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.hono</groupId>
-      <artifactId>service-base-test-utils</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.hono</groupId>
       <artifactId>core-test-utils</artifactId>
     </dependency>
     <dependency>
