--- test-utils/service-base-test-utils/pom.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright (c) 2020 Contributors to the Eclipse Foundation
-
-    See the NOTICE file(s) distributed with this work for additional
-    information regarding copyright ownership.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0
-
-    SPDX-License-Identifier: EPL-2.0
- -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.eclipse.hono</groupId>
-    <artifactId>test-utils</artifactId>
-    <version>1.7.0-SNAPSHOT</version>
-  </parent>
-  <artifactId>service-base-test-utils</artifactId>
-
-  <name>Hono Service Base Test Utils</name>
-  <description>Helper classes for implementing unit tests for components built on top of hono-service-base</description>
-  <url>https://www.eclipse.org/hono</url>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.eclipse.hono</groupId>
-      <artifactId>hono-service-base</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.hono</groupId>
-      <artifactId>core-test-utils</artifactId>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.assertj</groupId>
-      <artifactId>assertj-core</artifactId>
-      <scope>compile</scope>
-    </dependency>
-  </dependencies>
-</project>
